Split extension protocol request handlers.
